Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pymongo
faiss-cpu
sentence-transformers>=2.7.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
groq
google-generativeai
aiosmtplib
PyGithub
requests
pydantic[email]
email-validator
numpy
